Zone interaction work
WHAT WORKS:

Projectiles will now update their zone as they move appropriately.
Collisions with entities in the current zone of a projectile work.

ISSUES:

When projectile.update() is called in the game_update loop in app.js, or theprojectile updates in projectile.updateZone() in Projectile.js,

There will be a graphical error as the game updates the zone of the projectile constantly,